+++ app/code/Magento/Eav/Model/Entity/AbstractEntity.php
@@ -10,17 +10,18 @@ use Magento\Eav\Model\Entity\Attribute\AbstractAttribute;
 use Magento\Eav\Model\Entity\Attribute\Backend\AbstractBackend;
 use Magento\Eav\Model\Entity\Attribute\Frontend\AbstractFrontend;
 use Magento\Eav\Model\Entity\Attribute\Source\AbstractSource;
+use Magento\Eav\Model\Entity\Attribute\UniqueValidationInterface;
+use Magento\Eav\Model\ResourceModel\Attribute\DefaultEntityAttributes\ProviderInterface as DefaultAttributesProvider;
 use Magento\Framework\App\Config\Element;
+use Magento\Framework\App\ObjectManager;
 use Magento\Framework\DataObject;
 use Magento\Framework\DB\Adapter\DuplicateException;
 use Magento\Framework\Exception\AlreadyExistsException;
 use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Model\AbstractModel;
+use Magento\Framework\Model\ResourceModel\AbstractResource;
 use Magento\Framework\Model\ResourceModel\Db\ObjectRelationProcessor;
 use Magento\Framework\Model\ResourceModel\Db\TransactionManagerInterface;
-use Magento\Eav\Model\ResourceModel\Attribute\DefaultEntityAttributes\ProviderInterface as DefaultAttributesProvider;
-use Magento\Framework\Model\ResourceModel\AbstractResource;
-use Magento\Framework\App\ObjectManager;
 
 /**
  * Entity/Attribute/Model - entity abstract
@@ -216,11 +217,20 @@ abstract class AbstractEntity extends AbstractResource implements EntityInterfac
     protected $objectRelationProcessor;
 
     /**
+     * @var UniqueValidationInterface
+     */
+    private $uniqueValidator;
+
+    /**
      * @param Context $context
      * @param array $data
+     * @param UniqueValidationInterface|null $uniqueValidator
      */
-    public function __construct(Context $context, $data = [])
-    {
+    public function __construct(
+        Context $context,
+        $data = [],
+        UniqueValidationInterface $uniqueValidator = null
+    ) {
         $this->_eavConfig = $context->getEavConfig();
         $this->_resource = $context->getResource();
         $this->_attrSetEntity = $context->getAttributeSetEntity();
@@ -229,6 +239,8 @@ abstract class AbstractEntity extends AbstractResource implements EntityInterfac
         $this->_universalFactory = $context->getUniversalFactory();
         $this->transactionManager = $context->getTransactionManager();
         $this->objectRelationProcessor = $context->getObjectRelationProcessor();
+        $this->uniqueValidator = $uniqueValidator ?:
+            ObjectManager::getInstance()->get(UniqueValidationInterface::class);
         parent::__construct();
         $properties = get_object_vars($this);
         foreach ($data as $key => $value) {
@@ -400,61 +412,44 @@ abstract class AbstractEntity extends AbstractResource implements EntityInterfac
      *
      * @param string|int|Element $attribute
      * @return AbstractAttribute|false
+     * @throws LocalizedException
      * @SuppressWarnings(PHPMD.CyclomaticComplexity)
      */
     public function getAttribute($attribute)
     {
         /** @var $config \Magento\Eav\Model\Config */
         $config = $this->_getConfig();
-        if (is_numeric($attribute)) {
-            $attributeId = $attribute;
-            $attributeInstance = $config->getAttribute($this->getEntityType(), $attributeId);
-            if ($attributeInstance) {
-                $attributeCode = $attributeInstance->getAttributeCode();
-            }
-        } elseif (is_string($attribute)) {
-            $attributeCode = $attribute;
-            $attributeInstance = $config->getAttribute($this->getEntityType(), $attributeCode);
-            if (!$attributeInstance->getAttributeCode() && in_array($attribute, $this->getDefaultAttributes())) {
-                $attributeInstance->setAttributeCode(
-                    $attribute
-                )->setBackendType(
-                    AbstractAttribute::TYPE_STATIC
-                )->setIsGlobal(
-                    1
-                )->setEntity(
-                    $this
-                )->setEntityType(
-                    $this->getEntityType()
-                )->setEntityTypeId(
-                    $this->getEntityType()->getId()
-                );
-            }
-        } elseif ($attribute instanceof AbstractAttribute) {
-            $attributeInstance = $attribute;
-            $attributeCode = $attributeInstance->getAttributeCode();
+
+        $attributeInstance = $config->getAttribute($this->getEntityType(), $attribute);
+
+        if (!$attributeInstance->getAttributeCode() && in_array($attribute, $this->getDefaultAttributes(), true)) {
+            $attributeInstance = clone $attributeInstance;
+            $attributeInstance->setData([]);
+            $attributeInstance->setAttributeCode(
+                $attribute
+            )->setBackendType(
+                AbstractAttribute::TYPE_STATIC
+            )->setIsGlobal(
+                1
+            )->setEntity(
+                $this
+            )->setEntityType(
+                $this->getEntityType()
+            )->setEntityTypeId(
+                $this->getEntityType()->getId()
+            );
         }
 
-        if (empty($attributeInstance)
-            || !$attributeInstance instanceof AbstractAttribute
-            || !$attributeInstance->getId()
-            && !in_array($attributeInstance->getAttributeCode(), $this->getDefaultAttributes())
+        if (!$attributeInstance instanceof AbstractAttribute
+            || (!$attributeInstance->getId()
+                && !in_array($attributeInstance->getAttributeCode(), $this->getDefaultAttributes(), true))
         ) {
             return false;
         }
 
-        $attribute = $attributeInstance;
+        $this->addAttribute($attributeInstance);
 
-        if (!$attribute->getAttributeCode()) {
-            $attribute->setAttributeCode($attributeCode);
-        }
-        if (!$attribute->getAttributeModel()) {
-            $attribute->setAttributeModel($this->_getDefaultAttributeModel());
-        }
-
-        $this->addAttribute($attribute);
-
-        return $attribute;
+        return $attributeInstance;
     }
 
     /**
@@ -488,6 +483,7 @@ abstract class AbstractEntity extends AbstractResource implements EntityInterfac
     /**
      * Get attributes by scope
      *
+     * @param string $suffix
      * @return array
      */
     private function getAttributesByScope($suffix)
@@ -958,12 +954,8 @@ abstract class AbstractEntity extends AbstractResource implements EntityInterfac
 
         $data = $connection->fetchCol($select, $bind);
 
-        $objectId = $object->getData($entityIdField);
-        if ($objectId) {
-            if (isset($data[0])) {
-                return $data[0] == $objectId;
-            }
-            return true;
+        if ($object->getData($entityIdField)) {
+            return $this->uniqueValidator->validate($attribute, $object, $this, $entityIdField, $data);
         }
 
         return !count($data);
@@ -1674,14 +1666,16 @@ abstract class AbstractEntity extends AbstractResource implements EntityInterfac
         $connection->beginTransaction();
 
         try {
-            $select = $connection->select()->from($table, 'value_id')->where($where);
-            $origValueId = $connection->fetchOne($select);
+            $select = $connection->select()->from($table, ['value_id', 'value'])->where($where);
+            $origRow = $connection->fetchRow($select);
+            $origValueId = $origRow['value_id'] ?? false;
+            $origValue = $origRow['value'] ?? null;
 
             if ($origValueId === false && $newValue !== null) {
                 $this->_insertAttribute($object, $attribute, $newValue);
             } elseif ($origValueId !== false && $newValue !== null) {
                 $this->_updateAttribute($object, $attribute, $origValueId, $newValue);
-            } elseif ($origValueId !== false && $newValue === null) {
+            } elseif ($origValueId !== false && $newValue === null && $origValue !== null) {
                 $connection->delete($table, $where);
             }
             $this->_processAttributeValues();
@@ -1972,7 +1966,8 @@ abstract class AbstractEntity extends AbstractResource implements EntityInterfac
 
     /**
      * Load attributes for object
-     *  if the object will not pass all attributes for this entity type will be loaded
+     *
+     * If the object will not pass all attributes for this entity type will be loaded
      *
      * @param array $attributes
      * @param AbstractEntity|null $object
